Added inTriangle function
returns true if given point is in given triangle
